# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Spring PetClinic
  description: Spring PetClinic Sample Application.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: http://localhost:9966/petclinic/api
...
paths:
...
  /owners/{ownerId}/pets/{petId}/visits:
    post:
      tags:
        - owners
      operationId: addVisitToOwner
      summary: Adds a vet visit
      description: Records the details of a new vet visit.
      parameters:
        - name: ownerId
          in: path
          description: The ID of the pet owner.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 8
            example: 1
        - name: petId
          in: path
          description: The ID of the pet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 7
            example: 1
      requestBody:
        description: The details of the new vet visit.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitFields"
        required: true
      responses:
        201:
          description: The vet visit was successfully added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Pet not found for this owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /visits:
    get:
      tags:
        - visits
      operationId: listVisits
      summary: Lists visits
      description: Returns an array of visit .
      responses:
        200:
          description: visits found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
        - visits
      operationId: addVisit
      summary: Create a visit
      description: Creates a visit.
      requestBody:
        description: The visit
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitFields"
        required: true
      responses:
        200:
          description: visit created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Visit not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /visits/{visitId}:
    get:
      tags:
        - visits
      operationId: getVisit
      summary: Get a visit by ID
      description: Returns the visit or a 404 error.
      parameters:
        - name: visitId
          in: path
          description: The ID of the visit.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 2
            example: 1
      responses:
        200:
          description: Visit details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Visit not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    put:
      tags:
        - visits
      operationId: updateVisit
      summary: Update a visit by ID
      description: Returns the visit or a 404 error.
      parameters:
        - name: visitId
          in: path
          description: The ID of the visit.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 2
            example: 1
      requestBody:
        description: The visit
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitFields"
        required: true
      responses:
        204:
          description: Visit details found and returned.
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Visit not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      tags:
        - visits
      operationId: deleteVisit
      summary: Delete a visit by ID
      description: Returns the visit or a 404 error.
      parameters:
        - name: visitId
          in: path
          description: The ID of the visit.
          required: true
          schema:
            type: integer
            format: int32
            example: 4
      responses:
        200:
          description: Visit details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        404:
          description: Visit not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
    ProblemDetail:
      title: Problem Detail
      description: The schema for all error responses.
      type: object
      properties:
        type:
          title: Type
          description: Full URL that originated the error response.
          type: string
          format: uri
          example: "http://localhost:9966/petclinic/api/owner"
          readOnly: true
        title:
          title: Title
          description: The short error title.
          type: string
          example: NoResourceFoundException
          readOnly: true
        status:
          title: Status
          description: HTTP status code
          type: integer
          format: int32
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 500
          readOnly: true
        detail:
          title: Detail
          description: The long error message.
          type: string
          example: "No static resource api/owner."
          readOnly: true
        timestamp:
          title: Timestamp
          description: The time the error occurred.
          type: string
          format: date-time
          example: "2024-11-23T13:59:21.3820407Z"
          readOnly: true
        schemaValidationErrors:
          title: Schema validation errors
          description: Validation errors against the OpenAPI schema.
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
      required:
        - type
        - title
        - status
        - detail
        - timestamp
        - schemaValidationErrors
    ValidationMessage:
      title: Validation message
      description: Messages describing a validation error.
      type: object
      properties:
        message:
          title: Message
          description: The validation message.
          type: string
          example: "[Path '/lastName'] Instance type (null) does not match any allowed primitive type (allowed: ['string'])"
          readOnly: true
      required:
        - message
      additionalProperties: true
...
    VisitFields:
      title: Visit fields
      description: Editable fields of a vet visit.
      type: object
      properties:
        date:
          title: Date
          description: The date of the visit.
          type: string
          format: date
          example: "2013-01-01"
        description:
          title: Description
          description: The description for the visit.
          type: string
          minLength: 1
          maxLength: 255
          example: "rabies shot"
      required:
        - date
        - description
    Visit:
      title: Visit
      description: A booking for a vet visit.
      allOf:
        - $ref: "#/components/schemas/VisitFields"
        - type: object
          properties:
            id:
              title: ID
              description: The ID of the visit.
              type: integer
              format: int32
              minimum: 1
              maximum: 1000
              example: 1
              readOnly: true
            petId:
              title: Pet ID
              description: The ID of the pet.
              type: integer
              format: int32
              minimum: 1
              maximum: 1000
              example: 1
          required:
            - id
            - petId
...